0.0.12
---D3D12RHI---
Add BindlessDescriptors
Refactoring
-----------------
---Common---
Kill "Zombie" code
Refactor
Add custom cursor
----------------

+++ Dark Engine/Source/Runtime/Core/Private/Containers/DarkString.cpp
@@ -6,7 +6,7 @@
 
 FString::FString(const ANSICHAR* Str)
 {
-	int32 Lenght = strlen(Str) + 1;
+	int32 Lenght = (int32)strlen(Str) + 1;
 	_string.resize(Lenght);
 
 	size_t ConvertedChars = 0;
@@ -22,7 +22,7 @@ FString::FString(const std::wstring& Str)
 std::string FString::ToString() const
 {
 	std::string Result;
-	int32 Lenght = wcslen(_string.c_str()) + 1;
+	int32 Lenght = (int32)wcslen(_string.c_str()) + 1;
 	Result.resize(Lenght);
 
 	wcstombs_s(NULL, Result.data(), Lenght, _string.data(), Lenght - 1);
@@ -100,5 +100,5 @@ FString::FString(const WIDECHAR* Str)
 
 int32 FString::Len() const
 {
-	return _string.size();
+	return (int32)_string.size();
 }
